updated project to C#, moved Navigation and Dialog from Core to WPF project
