--- src/MagicMvvm.Core/Dialogs/IDialogParameters.cs
@@ -1,16 +1,0 @@
-﻿using MagicMvvm.Common;
-
-namespace MagicMvvm.Dialogs
-{
-    /// <summary>
-    /// Represents Dialog parameters.
-    /// </summary>
-    /// <remarks>
-    /// A class that implements this interface can be used to pass
-    /// object parameters during the showing and closing of Dialogs.
-    /// </remarks>
-    public interface IDialogParameters : IParameters
-    {
-
-    }
-}
