--- src/MagicMvvm.Core/Dialogs/IDialogResult.cs
@@ -1,19 +1,0 @@
-﻿namespace MagicMvvm.Dialogs
-{
-    /// <summary>
-    /// Contains <see cref="IDialogParameters"/> from the dialog
-    /// and the <see cref="ButtonResult"/> of the dialog.
-    /// </summary>
-    public interface IDialogResult
-    {
-        /// <summary>
-        /// The parameters from the dialog.
-        /// </summary>
-        IDialogParameters Parameters { get; }
-
-        /// <summary>
-        /// The result of the dialog.
-        /// </summary>
-        ButtonResult Result { get; }
-    }
-}
