+++ src/MagicMvvm.Core/IActiveAware.cs
@@ -1,22 +1,19 @@
-﻿using System;
+﻿namespace MagicMvvm;
 
-namespace MagicMvvm
+/// <summary>
+/// Interface that defines if the object instance is active
+/// and notifies when the activity changes.
+/// </summary>
+public interface IActiveAware
 {
     /// <summary>
-    /// Interface that defines if the object instance is active
-    /// and notifies when the activity changes.
+    /// Gets or sets a value indicating whether the object is active.
     /// </summary>
-    public interface IActiveAware
-    {
-        /// <summary>
-        /// Gets or sets a value indicating whether the object is active.
-        /// </summary>
-        /// <value><see langword="true" /> if the object is active; otherwise <see langword="false" />.</value>
-        bool IsActive { get; set; }
+    /// <value><see langword="true" /> if the object is active; otherwise <see langword="false" />.</value>
+    bool IsActive { get; set; }
 
-        /// <summary>
-        /// Notifies that the value for <see cref="IsActive"/> property has changed.
-        /// </summary>
-        event EventHandler IsActiveChanged;
-    }
+    /// <summary>
+    /// Notifies that the value for <see cref="IsActive"/> property has changed.
+    /// </summary>
+    event EventHandler IsActiveChanged;
 }
