+++ src/MagicMvvm.Core/Mvvm/BindableBase.cs
@@ -1,95 +1,91 @@
-﻿using System;
-using System.Collections.Generic;
-using System.ComponentModel;
-using System.Runtime.CompilerServices;
+﻿using System.Runtime.CompilerServices;
 
-namespace MagicMvvm
+namespace MagicMvvm;
+
+public abstract class BindableBase : INotifyPropertyChanged
 {
-    public abstract class BindableBase : INotifyPropertyChanged
+    /// <summary>
+    /// Occurs when a property value changes.
+    /// </summary>
+    public event PropertyChangedEventHandler PropertyChanged;
+
+    /// <summary>
+    /// Checks if a property already matches a desired value. Sets the property and
+    /// notifies listeners only when necessary.
+    /// </summary>
+    /// <typeparam name="T">Type of the property.</typeparam>
+    /// <param name="storage">Reference to a property with both getter and setter.</param>
+    /// <param name="value">Desired value for the property.</param>
+    /// <param name="propertyName">Name of the property used to notify listeners. This
+    /// value is optional and can be provided automatically when invoked from compilers that
+    /// support CallerMemberName.</param>
+    /// <returns>True if the value was changed, false if the existing value matched the
+    /// desired value.</returns>
+    protected virtual bool SetProperty<T>(ref T storage, T value, [CallerMemberName] string propertyName = null)
     {
-        /// <summary>
-        /// Occurs when a property value changes.
-        /// </summary>
-        public event PropertyChangedEventHandler PropertyChanged;
+        if (EqualityComparer<T>.Default.Equals(storage, value))
+            return false;
 
-        /// <summary>
-        /// Checks if a property already matches a desired value. Sets the property and
-        /// notifies listeners only when necessary.
-        /// </summary>
-        /// <typeparam name="T">Type of the property.</typeparam>
-        /// <param name="storage">Reference to a property with both getter and setter.</param>
-        /// <param name="value">Desired value for the property.</param>
-        /// <param name="propertyName">Name of the property used to notify listeners. This
-        /// value is optional and can be provided automatically when invoked from compilers that
-        /// support CallerMemberName.</param>
-        /// <returns>True if the value was changed, false if the existing value matched the
-        /// desired value.</returns>
-        protected virtual bool SetProperty<T>(ref T storage, T value, [CallerMemberName] string propertyName = null)
+        storage = value;
+        RaisePropertyChanged(propertyName);
+
+        return true;
+    }
+
+    /// <summary>
+    /// Checks if a property already matches a desired value. Sets the property and
+    /// notifies listeners only when necessary.
+    /// </summary>
+    /// <typeparam name="T">Type of the property.</typeparam>
+    /// <param name="storage">Reference to a property with both getter and setter.</param>
+    /// <param name="value">Desired value for the property.</param>
+    /// <param name="propertyName">Name of the property used to notify listeners. This
+    /// value is optional and can be provided automatically when invoked from compilers that
+    /// support CallerMemberName.</param>
+    /// <param name="onChanged">Action that is called after the property value has been changed.</param>
+    /// <returns>True if the value was changed, false if the existing value matched the
+    /// desired value.</returns>
+    protected virtual bool SetProperty<T>(ref T storage, T value, Action onChanged, [CallerMemberName] string propertyName = null)
+    {
+        if (EqualityComparer<T>.Default.Equals(storage, value))
+            return false;
+
+        storage = value;
+        onChanged?.Invoke();
+        RaisePropertyChanged(propertyName);
+
+        return true;
+    }
+
+    /// <summary>
+    /// Raises this object's PropertyChanged event.
+    /// </summary>
+    /// <param name="propertyName">Name of the property used to notify listeners. This
+    /// value is optional and can be provided automatically when invoked from compilers
+    /// that support <see cref="CallerMemberNameAttribute"/>.</param>
+    protected void RaisePropertyChanged([CallerMemberName] string propertyName = null)
+    {
+        OnPropertyChanged(new PropertyChangedEventArgs(propertyName));
+    }
+
+    /// <summary>
+    /// Raises all property's PropertyChanged event.
+    /// </summary>
+    /// <param name="propertiesNames">Name of properties</param>
+    protected void RaisePropertiesChanged(params string[] propertiesNames)
+    {
+        foreach (var propertyName in propertiesNames)
         {
-            if (EqualityComparer<T>.Default.Equals(storage, value))
-                return false;
-
-            storage = value;
             RaisePropertyChanged(propertyName);
-
-            return true;
         }
+    }
 
-        /// <summary>
-        /// Checks if a property already matches a desired value. Sets the property and
-        /// notifies listeners only when necessary.
-        /// </summary>
-        /// <typeparam name="T">Type of the property.</typeparam>
-        /// <param name="storage">Reference to a property with both getter and setter.</param>
-        /// <param name="value">Desired value for the property.</param>
-        /// <param name="propertyName">Name of the property used to notify listeners. This
-        /// value is optional and can be provided automatically when invoked from compilers that
-        /// support CallerMemberName.</param>
-        /// <param name="onChanged">Action that is called after the property value has been changed.</param>
-        /// <returns>True if the value was changed, false if the existing value matched the
-        /// desired value.</returns>
-        protected virtual bool SetProperty<T>(ref T storage, T value, Action onChanged, [CallerMemberName] string propertyName = null)
-        {
-            if (EqualityComparer<T>.Default.Equals(storage, value))
-                return false;
-
-            storage = value;
-            onChanged?.Invoke();
-            RaisePropertyChanged(propertyName);
-
-            return true;
-        }
-
-        /// <summary>
-        /// Raises this object's PropertyChanged event.
-        /// </summary>
-        /// <param name="propertyName">Name of the property used to notify listeners. This
-        /// value is optional and can be provided automatically when invoked from compilers
-        /// that support <see cref="CallerMemberNameAttribute"/>.</param>
-        protected void RaisePropertyChanged([CallerMemberName] string propertyName = null)
-        {
-            OnPropertyChanged(new PropertyChangedEventArgs(propertyName));
-        }
-
-        /// <summary>
-        /// Raises all property's PropertyChanged event.
-        /// </summary>
-        /// <param name="propertiesNames">Name of properties</param>
-        protected void RaisePropertiesChanged(params string[] propertiesNames)
-        {
-            foreach (var propertyName in propertiesNames)
-            {
-                RaisePropertyChanged(propertyName);
-            }
-        }
-
-        /// <summary>
-        /// Raises this object's PropertyChanged event.
-        /// </summary>
-        /// <param name="args">The PropertyChangedEventArgs</param>
-        protected virtual void OnPropertyChanged(PropertyChangedEventArgs args)
-        {
-            PropertyChanged?.Invoke(this, args);
-        }
+    /// <summary>
+    /// Raises this object's PropertyChanged event.
+    /// </summary>
+    /// <param name="args">The PropertyChangedEventArgs</param>
+    protected virtual void OnPropertyChanged(PropertyChangedEventArgs args)
+    {
+        PropertyChanged?.Invoke(this, args);
     }
 }
