--- src/MagicMvvm.Core/Navigation/INavigationParameters.cs
@@ -1,15 +1,0 @@
-﻿using MagicMvvm.Common;
-
-namespace MagicMvvm.Navigation
-{
-    /// <summary>
-    /// Represents Navigation parameters.
-    /// </summary>
-    /// <remarks>
-    /// This class can be used to to pass object parameters during Navigation.
-    /// </remarks>
-    public interface INavigationParameters : IParameters
-    {
-
-    }
-}
