--- src/MagicMvvm.Core/Navigation/NavigationParameters.cs
@@ -1,24 +1,0 @@
-﻿using MagicMvvm.Common;
-
-namespace MagicMvvm.Navigation
-{
-    /// <summary>
-    /// Represents Navigation parameters.
-    /// </summary>
-    /// <remarks>
-    /// This class can be used to to pass object parameters during Navigation.
-    /// </remarks>
-    public class NavigationParameters : ParametersBase, INavigationParameters
-    {
-        /// <summary>
-        /// Initializes a new instance of the <see cref="NavigationParameters"/> class.
-        /// </summary>
-        public NavigationParameters() : base() { }
-
-        /// <summary>
-        /// Constructs a list of parameters
-        /// </summary>
-        /// <param name="query">Query string to be parsed</param>
-        public NavigationParameters(string query) : base(query) { }
-    }
-}
