+++ src/MagicMvvm.Core/_Imports.cs
@@ -1,0 +1,6 @@
+﻿global using System;
+global using System.Collections;
+global using System.Collections.Generic;
+global using System.ComponentModel;
+global using System.Linq;
+global using System.Threading;
