+++ src/MagicMvvm.Wpf/Dialogs/DialogHost/DialogHostWindow.xaml
