+++ src/MagicMvvm.Wpf/_Imports.cs
@@ -1,0 +1,7 @@
+﻿global using System;
+global using System.Collections;
+global using System.Collections.Generic;
+global using System.Threading;
+global using System.Threading.Tasks;
+global using System.Windows;
+global using MagicMvvm.Common;
