--- tests/MagicMvvm.Core.Tests/Fixtures/DialogParametersFixture.cs
@@ -1,11 +1,0 @@
-﻿using MagicMvvm.Dialogs;
-
-namespace MagicMvvm.Core.Tests
-{
-    public class DialogParametersFixture : ParametersBaseFixture
-    {
-        public DialogParametersFixture() : base(new DialogParameters())
-        {
-        }
-    }
-}
