--- tests/MagicMvvm.Core.Tests/Fixtures/NavigationParametersFixture.cs
@@ -1,11 +1,0 @@
-﻿using MagicMvvm.Navigation;
-
-namespace MagicMvvm.Core.Tests
-{
-    public class NavigationParametersFixture : ParametersBaseFixture
-    {
-        public NavigationParametersFixture() : base(new NavigationParameters())
-        {
-        }
-    }
-}
