+++ tests/MagicMvvm.Core.Tests/MagicMvvm.Core.Tests.csproj
@@ -3,6 +3,8 @@
     <PropertyGroup>
         <TargetFramework>net5.0</TargetFramework>
         <IsPackable>false</IsPackable>
+        <LangVersion>10</LangVersion>
+        <PackageId>MagicMvvm.Core.Tests</PackageId>
     </PropertyGroup>
 
     <ItemGroup>
